modified according to the comments
