+++ federatedscope/core/data/utils.py
@@ -82,9 +82,7 @@ def load_dataset(config):
     elif config.data.type.lower() == 'vertical_fl_data':
         from federatedscope.vertical_fl.dataloader import load_vertical_data
         dataset, modified_config = load_vertical_data(config, generate=True)
-    elif config.data.type.lower() in [
-            'adult', 'abalone', 'givemesomecredit', 'blogfeedback'
-    ]:
+    elif config.data.type.lower() in ['adult', 'abalone', 'credit', 'blog']:
         from federatedscope.vertical_fl.dataloader import load_vertical_data
         dataset, modified_config = load_vertical_data(config, generate=False)
     elif 'movielens' in config.data.type.lower(
