+++ federatedscope/vertical_fl/dataloader/dataloader.py
@@ -2,9 +2,9 @@ import numpy as np
 
 from federatedscope.vertical_fl.dataset.adult import Adult
 from federatedscope.vertical_fl.dataset.abalone import Abalone
-from federatedscope.vertical_fl.dataset.give_me_some_credit\
-    import GiveMeSomeCredit
-from federatedscope.vertical_fl.dataset.blog_feedback import BlogFeedback
+from federatedscope.vertical_fl.dataset.credit \
+    import Credit
+from federatedscope.vertical_fl.dataset.blog import Blog
 
 
 def load_vertical_data(config=None, generate=False):
@@ -48,16 +48,16 @@ def load_vertical_data(config=None, generate=False):
                         algo=algo)
         data = dataset.data
         return data, config
-    elif name == 'givemesomecredit':
-        dataset = GiveMeSomeCredit(root=path,
-                                   name=name,
-                                   num_of_clients=config.federate.client_num,
-                                   feature_partition=feature_partition,
-                                   tr_frac=splits[0],
-                                   download=True,
-                                   seed=1234,
-                                   args=args,
-                                   algo=algo)
+    elif name == 'credit':
+        dataset = Credit(root=path,
+                         name=name,
+                         num_of_clients=config.federate.client_num,
+                         feature_partition=feature_partition,
+                         tr_frac=splits[0],
+                         download=True,
+                         seed=1234,
+                         args=args,
+                         algo=algo)
         data = dataset.data
         return data, config
     elif name == 'adult':
@@ -84,16 +84,16 @@ def load_vertical_data(config=None, generate=False):
                           algo=algo)
         data = dataset.data
         return data, config
-    elif name == 'blogfeedback':
-        dataset = BlogFeedback(root=path,
-                               name=name,
-                               num_of_clients=config.federate.client_num,
-                               feature_partition=feature_partition,
-                               tr_frac=splits[0],
-                               download=True,
-                               seed=1234,
-                               args=args,
-                               algo=algo)
+    elif name == 'blog':
+        dataset = Blog(root=path,
+                       name=name,
+                       num_of_clients=config.federate.client_num,
+                       feature_partition=feature_partition,
+                       tr_frac=splits[0],
+                       download=True,
+                       seed=1234,
+                       args=args,
+                       algo=algo)
         data = dataset.data
         return data, config
     elif generate:
