+++ federatedscope/vertical_fl/xgb_base/baseline/xgb_base_on_givemesomecredit.yaml
@@ -8,7 +8,7 @@ model:
   type: lr
 data:
   root: data/
-  type: givemesomecredit
+  type: credit
   batch_size: 2000
   splits: [0.8, 0.2]
 dataloader:
